FIX BUG: move list to tensor outside the torch.jit.script.

+++ lmcsc/model.py
@@ -99,7 +99,7 @@ class LMModel:
         self.set_convert_ids_to_tokens()
 
         self.tokenizer.padding_side = "left"
-        self.model.probs_template = torch.ones((self.model.vocab_size,)).to(
+        self.model.probs_template = torch.ones((self.model.vocab_size,), dtype=self.model.dtype).to(
             self.model.device
         )
